requirements.txt
streamlit>=1.28
python-dotenv
weaviate-client
opendatasets
pandas
